Massive refactor/restructuring of the ECS implementation, everything is now separated. Also did some minor refactoring
