+++ Habanero/engine/core/components/CameraComponent.h
@@ -5,12 +5,12 @@
 #ifndef HABANERO_CAMERACOMPONENT_H
 #define HABANERO_CAMERACOMPONENT_H
 
-#include "engine/core/ecs/ecs.h"
+#include "engine/core/ecs/component.h"
 
 /*
  * IMPORTANT INFO, rotation = {0.f, 0.f, 0.f} means that the camera is pointing towards negative z axis.
  */
-struct CameraComponent : ECS::Component {
+struct CameraComponent : ecs::Component {
     CameraComponent(float _fov, float _hRes, float _vRes, float _zNear, float _zFar)
             : m_fov(_fov), m_hRes(_hRes), m_vRes(_vRes), m_zNear(_zNear), m_zFar(_zFar) {}
