+++ Habanero/engine/core/ecs/entitymanager.cpp
@@ -1,23 +1,16 @@
-//
-// Created by notjustin on 2/9/18.
-//
-
-#include "ecs.h"
+#include "entitymanager.h"
 #include "EntityConstructor.h"
 
-using namespace ECS;
-
-// This one had to be in a .cpp file because of the forward declaration
-void Entity::addGroup(Group _group) noexcept {
+// This is to avoid a cyclical dependency issue and to work around the forward declared EntityManager in entity.h
+void ecs::Entity::addGroup(Group _group) noexcept {
     m_groupBitset[_group] = true;
     m_manager->addToGroup(this, _group);
 }
 
-void EntityManager::loadEntity(const std::string& _path) {
+void ecs::EntityManager::loadEntity(const std::string& _path) {
     EntityConstructor constructor;
     std::ifstream ifs(_path);
     const auto args = constructor.loadFile(ifs);
     constructor.constructEntity(args, nullptr);
     ifs.close();
 }
-
