+++ Habanero/engine/core/systems/CameraSystem.h
@@ -12,7 +12,7 @@
 #include "glm/matrix.hpp"
 #include "glm/gtc/matrix_transform.hpp"
 
-using namespace ECS;
+using namespace ecs;
 
 class CameraSystem : public System {
     friend class SystemManger;
@@ -24,10 +24,10 @@ public:
 #endif
     }
 
-	inline Entity* getActiveCamera() { return m_activeCamera; }
+	Entity* getActiveCamera() { return m_activeCamera; }
     inline void setActiveCamera(Entity *_cam) noexcept; // This essentially forces the system to use a specific camera,
                                                         // usually you'd have only one anyway.
-    glm::mat4 getMVPFromActiveCamera(glm::mat4 _modelMat);
+    glm::mat4 getMVPFromActiveCamera(glm::mat4 _modelMat) const;
 
 private:
